requirements.txt
requests
beautifulsoup4
python-dateutil
sqlalchemy
pdfplumber
python-docx
scikit-learn
pytest
# sentence-transformers # optional BERT

# Production-grade monitoring dependencies
aiohttp
tenacity
feedparser
pyyaml
kafka-python
python-dotenv
